add swapper-dump to bin, resue alien_provider
